requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
questionary>=2.0.0
pytesseract>=0.3.10
Pillow>=10.0.0